+++ build_runner/lib/src/io/asset_tracker.dart
@@ -7,10 +7,12 @@ import 'dart:io';
 
 import 'package:async/async.dart';
 import 'package:build/build.dart';
+import 'package:crypto/crypto.dart';
 import 'package:glob/glob.dart';
 import 'package:watcher/watcher.dart';
 
 import '../build/build_state/build_state.dart';
+import '../build/resolver/asset_ids.dart';
 import '../build_plan/build_configs.dart';
 import '../build_plan/build_packages.dart';
 import '../build_plan/build_step_plan.dart';
@@ -44,6 +46,7 @@ class AssetTracker {
       generatedSources,
       buildState,
       declaredAndActualOutputs,
+      buildStepPlan: buildStepPlan,
     );
   }
 
@@ -68,8 +71,9 @@ class AssetTracker {
     Set<AssetId> inputSources,
     Set<AssetId> generatedSources,
     BuildState buildState,
-    Iterable<AssetId> declaredAndActualOutputs,
-  ) async {
+    Iterable<AssetId> declaredAndActualOutputs, {
+    required BuildStepPlan buildStepPlan,
+  }) async {
     final allSources = <AssetId>{}
       ..addAll(inputSources)
       ..addAll(generatedSources);
@@ -97,8 +101,44 @@ class AssetTracker {
       final originalDigest = buildState.contentOfSource(id);
       if (originalDigest == null) continue;
 
-      final currentDigest = await _readerWriter.digest(id);
+      Digest? currentDigest;
+      try {
+        currentDigest = await _readerWriter.digest(id);
+      } on AssetNotFoundException {
+        continue;
+      } on FileSystemException {
+        continue;
+      }
       if (currentDigest != originalDigest.digest) {
+        updates[id] = ChangeType.MODIFY;
+      }
+    }
+
+    final preExistingOutputs = declaredAndActualOutputs.toSet().intersection(
+      allSources,
+    );
+    for (final id in preExistingOutputs) {
+      final originalContent = buildState.contentOf(
+        buildStepPlan: buildStepPlan,
+        id: id,
+      );
+      if (originalContent == null) continue;
+
+      Digest? currentDigest;
+      try {
+        currentDigest = await _readerWriter.digest(
+          id,
+          hidden: id.isHidden(
+            buildStepPlan: buildStepPlan,
+            buildState: buildState,
+          ),
+        );
+      } on AssetNotFoundException {
+        continue;
+      } on FileSystemException {
+        continue;
+      }
+      if (currentDigest != originalContent.digest) {
         updates[id] = ChangeType.MODIFY;
       }
     }